Refactoring: Tests and Mocks
