+++ __mocks__/build-subject.ts
@@ -1,6 +1,6 @@
 import { loadObservers } from "../lib/src";
-import { notifySalesDepartmentObservers } from "./observers/test-notify-sales-observers";
-import { sendEmailsObservers } from "./observers/test-send-email-observers";
+import { notifySalesDepartmentObservers } from "./notify-sales-observers";
+import { sendEmailsObservers } from "./send-email-observers";
 
 export const observers = loadObservers([
   ...sendEmailsObservers,
