+++ __tests__/Observer.test.ts
@@ -1,8 +1,8 @@
 import { observers } from "../__mocks__/build-subject";
-import { fakeUser } from "../__mocks__/fake-user";
-import { signUpUser, userRepo } from "../__mocks__/sign-up-user";
-import { mockNotifySalesDepartment } from "../__mocks__/observers/test-notify-sales-observers";
-import { mockSendEmails } from "../__mocks__/observers/test-send-email-observers";
+import { fakeUser } from "../__mocks__/use-case/fake-user";
+import { signUpUser, userRepo } from "../__mocks__/use-case/signup-user";
+import { mockNotifySalesDepartment } from "../__mocks__/notify-sales-observers";
+import { mockSendEmails } from "../__mocks__/send-email-observers";
 import { ObserverContainer } from "../lib/src/observers-container/entities/ObserverContainer";
 import { SubjectsMap } from "../lib/src/observers-container/entities/SubjectsMap";
 import { ObserversMap } from "../lib/src/observers-container/entities/ObserversMap";
